♻️ StudyMate에 entity 파일 생성
