+++ src/main/java/gaji/service/domain/room/converter/RoomConverter.java
@@ -2,7 +2,7 @@ package gaji.service.domain.room.converter;
 
 import gaji.service.domain.room.entity.RoomNotice;
 import gaji.service.domain.room.web.dto.RoomResponseDto;
-import gaji.service.domain.studyMate.Assignment;
+import gaji.service.domain.studyMate.entity.Assignment;
 
 public class RoomConverter {
