+++ src/main/java/gaji/service/domain/room/repository/AssignmentRepository.java
@@ -1,6 +1,6 @@
 package gaji.service.domain.room.repository;
 
-import gaji.service.domain.studyMate.Assignment;
+import gaji.service.domain.studyMate.entity.Assignment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
