+++ src/main/java/gaji/service/domain/room/repository/UserAssignmentRepository.java
@@ -1,7 +1,7 @@
 package gaji.service.domain.room.repository;
 
 import gaji.service.domain.room.entity.RoomEvent;
-import gaji.service.domain.studyMate.UserAssignment;
+import gaji.service.domain.studyMate.entity.UserAssignment;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
