+++ src/main/java/gaji/service/domain/room/service/RoomCommandServiceImpl.java
@@ -8,15 +8,13 @@ import gaji.service.domain.room.entity.NoticeConfirmation;
 import gaji.service.domain.room.entity.Room;
 import gaji.service.domain.room.entity.RoomEvent;
 import gaji.service.domain.room.entity.RoomNotice;
-import gaji.service.domain.room.repository.*;
 import gaji.service.domain.room.web.dto.RoomRequestDto;
-import gaji.service.domain.studyMate.Assignment;
-import gaji.service.domain.studyMate.StudyMate;
-import gaji.service.domain.studyMate.UserAssignment;
+import gaji.service.domain.studyMate.entity.Assignment;
+import gaji.service.domain.studyMate.entity.StudyMate;
+import gaji.service.domain.studyMate.entity.UserAssignment;
 import gaji.service.domain.studyMate.code.StudyMateErrorStatus;
 import gaji.service.domain.studyMate.repository.StudyMateRepository;
 import gaji.service.domain.studyMate.service.StudyMateQueryService;
-import gaji.service.domain.user.entity.User;
 import gaji.service.domain.user.service.UserQueryServiceImpl;
 import gaji.service.global.exception.RestApiException;
 import jakarta.transaction.Transactional;
