+++ src/main/java/gaji/service/domain/studyMate/entity/StudyApplicant.java
@@ -1,18 +1,17 @@
-package gaji.service.domain.studyMate;
+package gaji.service.domain.studyMate.entity;
 
 import gaji.service.domain.user.entity.User;
+import gaji.service.domain.enums.ApplicantStatus;
+import gaji.service.domain.room.entity.Room;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDateTime;
-
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ChatUser {
-
+public class StudyApplicant {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -21,12 +20,11 @@ public class ChatUser {
     @JoinColumn(name = "user_id")
     private User user;
 
-
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "chat_id")
-    private Chat chat;
+    @JoinColumn(name = "room_id")
+    private Room room;
 
-    //마지막 읽기 시점
-    private LocalDateTime lastReadTime;
-
+    //수락 대기 여부
+    @Enumerated(EnumType.STRING)
+    private ApplicantStatus status;
 }
