+++ src/main/java/gaji/service/domain/studyMate/entity/UserAssignment.java
@@ -1,4 +1,4 @@
-package gaji.service.domain.studyMate;
+package gaji.service.domain.studyMate.entity;
 
 import gaji.service.domain.user.entity.User;
 import jakarta.persistence.*;
