+++ src/main/java/gaji/service/domain/studyMate/service/StudyMateQueryService.java
@@ -1,6 +1,6 @@
 package gaji.service.domain.studyMate.service;
 
-import gaji.service.domain.studyMate.StudyMate;
+import gaji.service.domain.studyMate.entity.StudyMate;
 
 public interface StudyMateQueryService {
     StudyMate findByUserIdAndRoomId(Long id, Long roomId);
