+++ src/main/java/gaji/service/domain/studyMate/service/StudyMateQueryServiceImpl.java
@@ -1,6 +1,6 @@
 package gaji.service.domain.studyMate.service;
 
-import gaji.service.domain.studyMate.StudyMate;
+import gaji.service.domain.studyMate.entity.StudyMate;
 import gaji.service.domain.studyMate.code.StudyMateErrorStatus;
 import gaji.service.domain.studyMate.repository.StudyMateRepository;
 import gaji.service.global.exception.RestApiException;
